Add files via upload
Use Case 1: Financial Cash Flow

PROBLEM STATEMENT
An ABC company has released their financial data which has various data of accounts and also has the data of income and money paid to those accounts.
They want to create a financial dashboard with help of the data to analyze the finances of their company.
Some of the things expected in the dashboards are:
1)    Make Important KPIs
2)    Different Revenue through time
3)    Expenses by Account Type
4)    Cash flow statements for various months.
These are some of the expectations that are needed in the dashboard. You also have to add your own features to it.
The Dashboard should be clear enough to understand and should be attractive.
________________________________________________________________________________________________________________________________________________________________________________________

Use Case 2: Logistics

PROBLEM STATEMENT:
A logistics and supply chain company wants to make a dashboard for OTIF Analysis. 
On-time in-full (OTIF) is a supply chain metric for measuring performance in the logistics industry. OTIF generally refers to a supplier’s ability to deliver product within prescribed delivery windows and at full quantities ordered.
OTIF was designed to improve store operations within Walmart itself and quickly led to a series of major changes as it was quickly adopted by other retailers and companies.
So they have given data that considers 4 sheets. They are Orders, Salesperson, Customer and City. 
Expectations of Stakeholders are:-
1)	Make various KPIs.
2)	Explain the OTIF Performance over Time
3)	What is the Order delivery status over Time?
4)	What are the Highest OTIF Performances?
5)	Do something with Key Influencers and other advanced visualizations.
These are some of the features stakeholders want from their Data Analysts. You have to add your own creative features as well.
________________________________________________________________________________________________________________________________________________________________________________________

Use Case 3: Sales Target

PROBLEM STATEMENT:
An ABC company has the sales data for the last 3 years and they want to analyze it now. 
There is a Dimension Data where there is the list of companies and their Line of Business. There are 3 years of sales data. There is also a Target Data. 
Based on the data set given, try to create a dashboard.
Stakeholders want the following insights in the dashboard:
1)	Include important KPIs
2)	Sales over Time
3)	Sales for various Products
4)	Manager wise analysis
5)	Customer coverage across various Locations
6)	Target Analysis
These are some of the points stakeholders needed. Also, add your own insights to it.